Split Test Cases
In order to prevent confusion I split my test cases into two separate classes.
