+++ LCATest.java
@@ -5,14 +5,6 @@ import org.junit.Test;
 public class LCATest {
 	
 	
-	
-	@Test
-	public void testDAGLCA() {
-		
-		
-		// Need to write a test for testing LCA for a Directed Acyclic Graph.
-		// As of yet I am not sure what format this test should take on.
-	}
 
 	@Test
 	public void testLCA() {
@@ -202,52 +194,6 @@ public class LCATest {
 		assertEquals("Testing contains", true, LCA.contains(7));
 	}
 	
-	//Following tests check Directed Graph class works correctly
-		@Test
-		public void testForDirectedGraph(){
-			DAG test = new DAG(10);
-			test.addEdge(1, 2);
-			test.addEdge(1, 3);
-			test.addEdge(3, 4);
-			test.addEdge(4, 5);
-			test.addEdge(4, 6);
-			
-			assertEquals("", 1, test.indegree(4));
-			assertEquals("", 2, test.outdegree(4));
-			assertEquals("", 5, test.E());
-			assertEquals("", 10, test.V());
-			String ans = "[5, 6]";
-			assertEquals("",ans, test.adj(4).toString());
-		}
-		
-		
-		@Test
-		public void testAddEdge(){
-			DAG test4 = new DAG(5);
-			test4.addEdge(3, 9);
-			test4.addEdge(-2, -5);	
-		}
-		
-		@Test
-		public void testinDegree(){
-			DAG test5 = new DAG(5);
-			assertEquals("", -1, test5.indegree(-3));
-		}
-		
-		@Test
-		public void testOutDegree(){
-			DAG test6 = new DAG(5);
-			assertEquals("", -1, test6.outdegree(8));	
-		}
-		
-		
-		@Test(expected=Exception.class)
-		public void exceptionTest(){
-			//Can't make a directed graph with less than 0 vertices
-			DAG test3 = new DAG(-5);
-		}
-		
-
